requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
scikit-learn
matplotlib
seaborn
segment-anything
